update user func

+++ src/controllers/season.controller.js
@@ -2,40 +2,58 @@ const SeasonService = require('../services/season.service');
 
 const SeasonController = {}
 
-SeasonController.getAnimeSeason = async (req, res, next) => {
-    try {
-        const id = req.params.id;
-        const rs = await SeasonService.getAnimeSeason(id);
-        return res.status(200).json(rs);
-    } catch (error) {
-        console.log(error);
-        return res.status(500).json({status: "Error", message: "Has a fucking error"});
+SeasonController.getAnimeSeason = (req, res, next) => {
+    const id = req.params.id;
+    if(id){
+        SeasonService.getAnimeSeason(id)
+        .then(rs => {
+            return res.status(200).json(rs);
+        })
+        .catch(err => {
+            console.log(err);
+            return res.status(500).json({status: "Error", message: "Has a fucking error"});
+        });
+    }else{
+        return res.status(400).json({status: "Failed", message: "Missing params"});
     }
 }
 
 SeasonController.addSeason = async (req, res, next) => {
-    try {
+    const idAnime = req.body.idAnime;
+    const idSeries = req.body.idSeries;
+    const season = req.body.season;
+    if(idAnime && idSeries && season){
         const data = {
-            idAnime: req.body.idAnime,
-            idSeries: req.body.idSeries,
-            season: req.body.season
+            idAnime,
+            idSeries,
+            season
         }
-        const rs = await SeasonService.addSeason(data);
-        return res.status(200).json(rs);
-    } catch (error) {
-        console.log(error);
-        return res.status(500).json({status: "Error", message: "Has a fucking error"});
+        SeasonService.addSeason(data)
+        .then(rs => {
+            return res.status(200).json(rs);
+        })
+        .catch(err => {
+            console.log(err);
+            return res.status(500).json({status: "Error", message: "Has a fucking error"});
+        });
+    }else{
+        return res.status(400).json({status: "Failed", message: "Missing params"});
     }
 }
 
 SeasonController.deleteSeason = async (req, res, next) => {
-    try {
-        const id = req.params.id;
-        const rs = await SeasonService.deleteSeason(id);
-        return res.status(200).json(rs);
-    } catch (error) {
-        console.log(error);
+    const id =  req.params.id;
+    if(id){
+        SeasonService.deleteSeason(id)
+        .then(rs => {
+            return res.status(200).json(rs);
+        })
+        .catch(err => {
+            console.log(err);
         return res.status(500).json({status: "Error", message: "Has a fucking error"});
+        })
+    }else{
+        return res.status(400).json({status: "Failed", message: "Missing params"});
     }
 }
