+++ src/controllers/status.controller.js
@@ -14,42 +14,59 @@ StatusController.getAll = (req, res , next) => {
 }
 
 StatusController.createOne = (req, res , next) => {
-    const data = {status: req.body.status};
-    StatusService.createOne(data)
-    .then(rs => {
-        return res.status(200).json(rs);
-    })
-    .catch(err => {
-        console.log(err);
-        return res.status(500).json({status: "Error", message: "Has a fucking error"});
-    })
+    const status = req.body.status;
+    if(status){
+        const data = {status};
+        StatusService.createOne(data)
+        .then(rs => {
+            return res.status(200).json(rs);
+        })
+        .catch(err => {
+            console.log(err);
+            return res.status(500).json({status: "Error", message: "Has a fucking error"});
+        })
+    }else{
+        return res.status(400).json({status: "Failed", message: "Missing params"});
+    }
+    
 }
 
 StatusController.updateOne = (req, res , next) => {
-    const data = {
-        id: req.params.id,
-        status: req.body.status
+    const id = req.params.id;
+    const status = req.body.status;
+    if(id && status){
+        const data = {
+            id,
+            status
+        }
+        StatusService.updateOne(data)
+        .then(rs => {
+            return res.status(200).json(rs);
+        })
+        .catch(err => {
+            console.log(err);
+            return res.status(500).json({status: "Error", message: "Has a fucking error"});
+        })
+    }else{
+        return res.status(400).json({status: "Failed", message: "Missing params"});
     }
-    StatusService.updateOne(data)
-    .then(rs => {
-        return res.status(200).json(rs);
-    })
-    .catch(err => {
-        console.log(err);
-        return res.status(500).json({status: "Error", message: "Has a fucking error"});
-    })
+    
 }
 
 StatusController.deleteOne = (req, res , next) => {
     const id = req.params.id
-    StatusService.deleteOne(id)
-    .then(rs => {
-        return res.status(200).json(rs);
-    })
-    .catch(err => {
-        console.log(err);
-        return res.status(500).json({status: "Error", message: "Has a fucking error"});
-    })
+    if(id){
+        StatusService.deleteOne(id)
+        .then(rs => {
+            return res.status(200).json(rs);
+        })
+        .catch(err => {
+            console.log(err);
+            return res.status(500).json({status: "Error", message: "Has a fucking error"});
+        })
+    }else{
+        return res.status(400).json({status: "Failed", message: "Missing params"});
+    }
 }
 
 module.exports = StatusController;
